backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "distro==1.9.0",
    "faiss-cpu==1.11.0",
    "fastapi==0.115.12",
    "flask==3.1.1",
    "google-api-core==2.24.2",
    "google-api-python-client==2.170.0",
    "google-auth==2.40.2",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.2",
    "googleapis-common-protos==1.70.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "lxml==5.4.0",
    "markupsafe==3.0.2",
    "numpy==2.2.6",
    "oauthlib==3.2.2",
    "openai==1.82.0",
    "packaging==25.0",
    "proto-plus==1.26.1",
    "protobuf==6.31.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pyparsing==3.2.3",
    "pypdf2==3.0.1",
    "python-docx==1.1.2",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "starlette==0.46.2",
    "tqdm==4.67.1",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.1",
    "uritemplate==4.1.1",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "werkzeug==3.1.3",
]
